Integrate xcode.rs into platform/macos.rs
